Update gradient styles
